fix: iana subtag language lookup optimisations
- previous solution loaded all 8138 options into a list by default then
  ran a membership check (O(n)) to search for each language.
- optimisations:
  - if the language name is "default", or is too short to match the
    language code regex, then skip executing the regex.
  - read the code list data at most once, only if needed.
  - put the code strings into a set for faster membership check (O(1)).
  - split the code list into the shorter list of 2-character codes
    (e.g. en, fr, de), and check that first. Assuming these shorter
    codes are more likely to be used, it is faster to check and uses
    less memory than loading the full list (8KB short vs 525KB).
- best case: default, invalid, or short language codes get faster
  lookup with less memory used than before.
- worst case: longer language codes get faster lookup with the same
  memory used as before.
